pyproject.toml
[project]
name = "pythonfy"
version = "0.1.0"
description = ""
authors = [
    {name = "joao pedro",email = "joaopedrog0667@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "spotipy (>=2.25.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "flask (>=3.1.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
